star_wars/requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.7
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.12.4
idna==2.10
ipdb==0.13.7
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.8.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
